chore(weave): move client/server-shared leaf modules to weave/shared
The client needs ids, constants, client-facing exception types, feedback type
conventions, and the ext/int ref converter — none of which are server
implementation. Move them into weave/shared (the established home for code
shared between the client and the trace server) so client code can stop
importing weave.trace_server:

- weave/trace_server/ids.py                      -> weave/shared/ids.py
- weave/trace_server/constants.py                -> weave/shared/constants.py
- weave/trace_server/interface/feedback_types.py -> weave/shared/feedback_types.py
- weave/trace_server/trace_server_converter.py   -> weave/shared/trace_server_converter.py
- exception classes from weave/trace_server/errors.py -> weave/shared/errors.py
  (the server-side error registry / HTTP status mapping stays put and
  re-exports the classes, so exception identity is preserved)

Every old import path keeps working via shims; server-side imports are left
untouched on purpose. Client modules (weave/trace, weave/flow, weave/chat,
weave/evaluation, weave/integrations, weave/utils, weave/durability,
weave/trace_server_bindings) now import from weave.shared.
