+++ weave/evaluation/otel_eval_linker.py
@@ -25,7 +25,7 @@ from weave.evaluation.eval import (
     _find_current_evaluate_call,
     _find_current_predict_and_score_call,
 )
-from weave.trace_server import constants
+from weave.shared import constants
 from weave.trace_server import trace_server_interface as tsi
 
 if TYPE_CHECKING:
