+++ weave/trace/call.py
@@ -7,6 +7,7 @@ from collections.abc import Callable
 from concurrent.futures import Future
 from typing import TYPE_CHECKING, Any, TypedDict
 
+from weave.shared.constants import MAX_DISPLAY_NAME_LENGTH
 from weave.trace import urls
 from weave.trace.context import weave_client_context
 from weave.trace.feedback import RefFeedbackQuery
@@ -18,7 +19,6 @@ from weave.trace.serialization.serialize import from_json
 from weave.trace.util import log_once
 from weave.trace.vals import WeaveObject
 from weave.trace_server.common_interface import SortBy
-from weave.trace_server.constants import MAX_DISPLAY_NAME_LENGTH
 from weave.trace_server.interface.query import Query
 from weave.trace_server.trace_server_interface import (
     CallSchema,
