+++ weave/trace/project_id_resolver.py
@@ -15,8 +15,8 @@ from typing import TYPE_CHECKING, Any
 
 from httpx import HTTPStatusError as HTTPError
 
+from weave.shared.errors import DigestMismatchError
 from weave.trace.settings import should_enable_client_side_digests
-from weave.trace_server.errors import DigestMismatchError
 from weave.trace_server.trace_server_interface import ProjectsInfoReq
 
 if TYPE_CHECKING:
