+++ weave/trace/vals.py
@@ -8,6 +8,7 @@ from typing import Any, Literal, SupportsIndex, cast
 
 from pydantic import BaseModel
 
+from weave.shared.errors import ObjectDeletedError
 from weave.trace import box
 from weave.trace.context.tests_context import get_raise_on_captured_errors
 from weave.trace.context.weave_client_context import get_weave_client
@@ -25,7 +26,6 @@ from weave.trace.refs import (
 )
 from weave.trace.serialization.serialize import from_json
 from weave.trace.table import Table
-from weave.trace_server.errors import ObjectDeletedError
 from weave.trace_server.trace_server_interface import (
     ObjReadReq,
     TableQueryReq,
