+++ weave/trace/weave_client.py
@@ -19,12 +19,21 @@ from httpx import HTTPStatusError as HTTPError
 from weave.chat.chat import Chat
 from weave.chat.inference_models import InferenceModels
 from weave.durability.wal_manager import WALManager
+from weave.shared.constants import MAX_OBJECT_NAME_LENGTH
 from weave.shared.digest import (
     compute_file_digest,
     compute_object_digest,
     compute_row_digest,
     compute_table_digest,
 )
+from weave.shared.errors import DigestMismatchError, InvalidExternalRef
+from weave.shared.feedback_types import (
+    RUNNABLE_FEEDBACK_TYPE_PREFIX,
+    runnable_feedback_output_selector,
+    runnable_feedback_runnable_ref_selector,
+)
+from weave.shared.ids import generate_id
+from weave.shared.trace_server_converter import universal_ext_to_int_ref_converter
 from weave.telemetry import trace_sentry
 from weave.trace import settings
 from weave.trace.call import (
@@ -107,15 +116,6 @@ from weave.trace.wandb_run_context import (
 )
 from weave.trace.weave_client_send_file_cache import WeaveClientSendFileCache
 from weave.trace_server.common_interface import AnnotationQueueItemsFilter, SortBy
-from weave.trace_server.constants import MAX_OBJECT_NAME_LENGTH
-from weave.trace_server.errors import DigestMismatchError, InvalidExternalRef
-from weave.trace_server.ids import generate_id
-from weave.trace_server.interface.feedback_types import (
-    RUNNABLE_FEEDBACK_TYPE_PREFIX,
-    runnable_feedback_output_selector,
-    runnable_feedback_runnable_ref_selector,
-)
-from weave.trace_server.trace_server_converter import universal_ext_to_int_ref_converter
 from weave.trace_server.trace_server_interface import (
     AliasesListReq,
     AnnotationQueueAddCallsReq,
