+++ weave/trace_server_bindings/stainless_remote_http_trace_server.py
@@ -48,6 +48,7 @@ from typing_extensions import Self
 from weave_server_sdk import WeaveTrace
 from weave_server_sdk import models as sdk_models
 
+from weave.shared.ids import generate_id
 from weave.trace.env import ssl_verify, weave_trace_server_url
 from weave.trace.settings import (
     http_timeout,
@@ -56,7 +57,6 @@ from weave.trace.settings import (
     should_use_calls_complete,
 )
 from weave.trace_server import trace_server_interface as tsi
-from weave.trace_server.ids import generate_id
 from weave.trace_server.service_interface import ServerInfoRes
 from weave.trace_server_bindings.async_batch_processor import AsyncBatchProcessor
 from weave.trace_server_bindings.call_batch_processor import CallBatchProcessor
